Improve eslint situation
ESLint was complaining about a bunch of things in tests, like undefined
variables that are global and some other small things. This commit fixes
most of these issues. I also modified the tool so it runs on more code
by default. This should give much better coverage for this project.

As a result, I spotted a couple of files that look suspiciously
erroneous that I ended up removing here. They seem to have appeared in a
merge commit: 091f339b175.