Expose generateFile as named export from generator module
It's possible (and appealing) to use Webpack's
[`val-loader`](https://github.com/webpack-contrib/val-loader) to
generate the legacy browser `prefixData` module on the fly at
compilation time; however, this requires the ability to generate the
code synchronously and capture it as a string.

We can enable this by merely exposing the `generateFile` method as a
named export from the `generator` module.